music-discovery-database/requirements.txt
requests==2.31.0  # HTTP library for web scraping
beautifulsoup4==4.12.2  # HTML parsing library
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
pytest==7.4.3  # Testing framework
